Création des droits d'accès avec Pundit

+++ app/controllers/clients_controller.rb
@@ -2,20 +2,24 @@ class ClientsController < ApplicationController
   before_action :set_client, only: [ :show, :edit, :update, :destroy ]
 
   def index
+    authorize Client
     @q = params[:q].to_s.strip
-    scope = Client.order(:legal_name).search(@q)
+    scope = policy_scope(Client).order(:legal_name).search(@q)
     @clients = paginate(scope)
   end
 
   def show
+    authorize @client
   end
 
   def new
     @client = Client.new
+    authorize @client
   end
 
   def create
     @client = Client.new(client_params)
+    authorize @client
 
     if @client.save
       redirect_to @client, notice: "Client cree avec succes."
@@ -25,9 +29,12 @@ class ClientsController < ApplicationController
   end
 
   def edit
+    authorize @client
   end
 
   def update
+    authorize @client
+
     if @client.update(client_params)
       redirect_to @client, notice: "Client mis a jour avec succes."
     else
@@ -36,6 +43,8 @@ class ClientsController < ApplicationController
   end
 
   def destroy
+    authorize @client
+
     if @client.destroy
       redirect_to clients_path, status: :see_other, notice: "Client supprime avec succes."
     else
