+++ app/controllers/placements_controller.rb
@@ -3,21 +3,25 @@ class PlacementsController < ApplicationController
   before_action :set_form_collections, only: [ :new, :create, :edit, :update ]
 
   def index
+    authorize Placement
     @q = params[:q].to_s.strip
     @status = params[:status].to_s.strip
-    scope = Placement.includes(:mission, :candidate).order(created_at: :desc).search(@q).with_status(@status)
+    scope = policy_scope(Placement).includes(:mission, :candidate).order(created_at: :desc).search(@q).with_status(@status)
     @placements = paginate(scope)
   end
 
   def show
+    authorize @placement
   end
 
   def new
     @placement = Placement.new
+    authorize @placement
   end
 
   def create
     @placement = Placement.new(placement_params)
+    authorize @placement
 
     if @placement.save
       redirect_to @placement, notice: "Placement cree avec succes."
@@ -27,9 +31,12 @@ class PlacementsController < ApplicationController
   end
 
   def edit
+    authorize @placement
   end
 
   def update
+    authorize @placement
+
     if @placement.update(placement_params)
       redirect_to @placement, notice: "Placement mis a jour avec succes."
     else
@@ -38,6 +45,8 @@ class PlacementsController < ApplicationController
   end
 
   def destroy
+    authorize @placement
+
     if @placement.destroy
       redirect_to placements_path, status: :see_other, notice: "Placement supprime avec succes."
     else
@@ -52,8 +61,8 @@ class PlacementsController < ApplicationController
   end
 
   def set_form_collections
-    @missions = Mission.order(:reference)
-    @candidates = Candidate.order(:last_name, :first_name)
+    @missions = policy_scope(Mission).order(:reference)
+    @candidates = policy_scope(Candidate).order(:last_name, :first_name)
   end
 
   def placement_params
